RedshiftAnalyticsService/src/requirements.txt
flask
psycopg2-binary
pymongo